requirements.txt
absl-py==0.15.0
aiofile==3.7.2
aiofiles==0.7.0
anyio==3.3.4
asgiref==3.4.1
astunparse==1.6.3
cachetools==4.2.4
caio==0.9.3
certifi==2021.10.8
charset-normalizer==2.0.7
click==7.1.2
colorama==0.4.4
cycler==0.10.0
fastapi==0.70.0
flatbuffers==1.12
gast==0.3.3
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h11==0.12.0
h5py==2.10.0
idna==3.3
imageai==2.1.6
imageio==2.9.0
importlib-metadata==4.8.1
Keras==2.4.3
Keras-Preprocessing==1.1.2
keras-resnet==0.2.0
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.3.2
numpy==1.19.3
oauthlib==3.1.1
opencv-python==4.5.4.58
opt-einsum==3.3.0
Pillow==7.0.0
protobuf==3.19.0
pur==5.4.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pyparsing==2.4.7
python-dateutil==2.8.2
python-multipart==0.0.5
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.4.1
six==1.15.0
sniffio==1.2.0
starlette==0.16.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.7
uvicorn==0.15.0
Werkzeug==2.0.2
wrapt==1.12.1
zipp==3.6.0
